fix: extract SQLiteCache and CachedResponse to shared providers/cache.py
- Extracted `SQLiteCache` and `CachedResponse` to `src/chessclub/providers/cache.py`
- Updated `src/chessclub/providers/chesscom/cache.py` to re-export for backward compatibility
- Updated imports in Chess.com provider, Lichess provider, and CLI
- Updated docs/cache.md to reflect new location
- Removed TODO in Lichess client
